requirements.txt
aiohttp
termcolor
ruamel.yaml
python-slugify
